fix data type

+++ app/schemas/telemetry.py
@@ -25,7 +25,7 @@ class DeviceResponse(BaseModel):
     created_at: datetime
     group_id: Optional[int] = None
     group_name: Optional[str] = None
-    moisture_lvl: Optional[int] = None
-    battery_lvl: Optional[int] = None
-    water_lvl: Optional[int] = None
+    moisture_lvl: Optional[float] = None
+    battery_lvl: Optional[float] = None
+    water_lvl: Optional[float] = None
     last_seen: Optional[datetime] = None
